# node-api/swagger.yaml
openapi: 3.0.0
info:
  title: Node.js Sign-up and Verification API
  description: Node.js + MySQL - API with email sign-up, verification, authentication and forgot password
  version: 1.0.0
servers:
  - url: http://192.168.50.50:8002
    description: Local development server
paths:
...
  /organizations:
            get:
              summary: Get a list of all Organizations
              description: Restricted to admin users.
              operationId: getAllOrganizations
              security:
                - bearerAuth: []
              responses:
                "200":
                  description: An array of all accounts
                  content:
                    application/json:
                      schema:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              example: "5eb12e197e06a76ccdefc121"
                            name:
                              type: string
                              example: "Company name"
                            website:
                              type: string
                              example: "test.com"
                            created:
                              type: string
                              example: "2020-05-05T09:12:57.848Z"
                            updated:
                              type: string
                              example: "2020-05-08T03:11:21.553Z"
                "401":
                  $ref: "#/components/responses/UnauthorizedError"
            post:
              summary: Create a new Organization
              description: Restricted to admin users.
              operationId: createOrganization
              security:
                - bearerAuth: []
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      type: object
                      properties:
                        name:
                          type: string
                          example: "Company name"
                        website:
                          type: string
                          example: "test.com"
                      required:
                        - name
                        - website
              responses:
                "200":
                  description: Organization created successfully, verification is not required for Organization created with this endpoint. The details of the new Organization are returned.
                  content:
                    application/json:
                      schema:
                        type: object
                        properties:
                          id:
                            type: string
                            example: "5eb12e197e06a76ccdefc121"
                          name:
                            type: string
                            example: "Company name"
                          website:
                            type: string
                            example: "test.com"
                          created:
                            type: string
                            example: "2020-05-05T09:12:57.848Z"
                "400":
                  description: Organization is already registered
                  content:
                    application/json:
                      schema:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Organization already registered"
                "401":
                  $ref: "#/components/responses/UnauthorizedError"
  /organizations/{id}:
            parameters:
              - in: path
                name: id
                description: Organization id
                required: true
                example: "5eb12e197e06a76ccdefc121"
                schema:
                  type: string
            get:
              summary: Get a single Organization by id
              description: Admin users can access any Organization, regular users are restricted to their own organization.
              operationId: getOrganizationById
              security:
                - bearerAuth: []
              responses:
                "200":
                  description: Details of the specified Organization
                  content:
                    application/json:
                      schema:
                        type: object
                        properties:
                          id:
                            type: string
                            example: "5eb12e197e06a76ccdefc121"
                          name:
                            type: string
                            example: "Organization name"
                          website:
                            type: string
                            example: "test.com"
                          created:
                            type: string
                            example: "2020-05-05T09:12:57.848Z"
                          updated:
                            type: string
                            example: "2020-05-08T03:11:21.553Z"
                "404":
                  $ref: "#/components/responses/NotFoundError"
                "401":
                  $ref: "#/components/responses/UnauthorizedError"
            put:
              summary: Update an Organization
              description: Admin users can update any Organization including role, regular users are restricted to their own Organization and cannot update role.
              operationId: updateOrganization
              security:
                - bearerAuth: []
              requestBody:
                required: true
                content:
                  application/json:
                    schema:
                      type: object
                      properties:
                        name:
                          type: string
                          example: "Organization name"
                        website:
                          type: string
                          example: "test.com"
              responses:
                "200":
                  description: Organization updated successfully. The details of the updated Organization are returned.
                  content:
                    application/json:
                      schema:
                        type: object
                        properties:
                          id:
                            type: string
                            example: "5eb12e197e06a76ccdefc121"
                          name:
                            type: string
                            example: "Organization name"
                          website:
                            type: string
                            example: "test.com"
                          created:
                            type: string
                            example: "2020-05-05T09:12:57.848Z"
                          updated:
                            type: string
                            example: "2020-05-08T03:11:21.553Z"
                "404":
                  $ref: "#/components/responses/NotFoundError"
                "401":
                  $ref: "#/components/responses/UnauthorizedError"
            delete:
              summary: Delete an Organization
              description: Admin users can delete any Organization, regular users are restricted to their own Organization.
              operationId: deleteOrganization
              security:
                - bearerAuth: []
              responses:
                "200":
                  description: Organization deleted successfully
                  content:
                    application/json:
                      schema:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Organization deleted successfully"
                "404":
                  $ref: "#/components/responses/NotFoundError"
                "401":
                  $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Not Found"